Lag egne DataGenerator for hver klasse
