+++ src/test/kotlin/no/nav/familie/ba/sak/dataGenerator/brev/BrevPeriodeGrunnlagMedPersoner.kt
@@ -1,13 +1,9 @@
-package no.nav.familie.ba.sak.dataGenerator
+package no.nav.familie.ba.sak.dataGenerator.brev
 
-import no.nav.familie.ba.sak.common.lagTriggesAv
-import no.nav.familie.ba.sak.kjerne.vedtak.begrunnelser.TriggesAv
-import no.nav.familie.ba.sak.kjerne.vedtak.begrunnelser.VedtakBegrunnelseSpesifikasjon
-import no.nav.familie.ba.sak.kjerne.vedtak.begrunnelser.VedtakBegrunnelseType
-import no.nav.familie.ba.sak.kjerne.vedtak.vedtaksperiode.Vedtaksperiodetype
 import no.nav.familie.ba.sak.kjerne.brev.domene.BrevBegrunnelseGrunnlagMedPersoner
 import no.nav.familie.ba.sak.kjerne.brev.domene.BrevPeriodeGrunnlagMedPersoner
 import no.nav.familie.ba.sak.kjerne.brev.domene.MinimertUtbetalingsperiodeDetalj
+import no.nav.familie.ba.sak.kjerne.vedtak.vedtaksperiode.Vedtaksperiodetype
 import java.time.LocalDate
 
 fun lagBrevPeriodeGrunnlagMedPersoner(
@@ -27,19 +23,5 @@ fun lagBrevPeriodeGrunnlagMedPersoner(
         fritekster = fritekster,
         minimerteUtbetalingsperiodeDetaljer = minimertUtbetalingsperiodeDetalj,
         erFørsteVedtaksperiodePåFagsak = erFørsteVedtaksperiodePåFagsak
-    )
-}
-
-fun lagBrevBegrunnelseGrunnlagMedPersoner(
-    vedtakBegrunnelseSpesifikasjon: VedtakBegrunnelseSpesifikasjon = VedtakBegrunnelseSpesifikasjon.INNVILGET_BOSATT_I_RIKTET,
-    vedtakBegrunnelseType: VedtakBegrunnelseType = VedtakBegrunnelseType.INNVILGET,
-    triggesAv: TriggesAv = lagTriggesAv(),
-    personIdenter: List<String> = emptyList(),
-): BrevBegrunnelseGrunnlagMedPersoner {
-    return BrevBegrunnelseGrunnlagMedPersoner(
-        vedtakBegrunnelseSpesifikasjon = vedtakBegrunnelseSpesifikasjon,
-        vedtakBegrunnelseType = vedtakBegrunnelseType,
-        triggesAv = triggesAv,
-        personIdenter = personIdenter
     )
 }
